Provide better error output when given an invalid SQLite file
If given a file that does not exist this will create it.  But we do
not want to delete the file in that case because someone could
accidentally give the program a legitimate file, which we do not want
to delete.  It’s better to create an empty file by mistake than assume
we can delete a file.