Teach build_search_query() to accept a list
Right now build_search_query() uses the global search_terms list.
This patch changes it to accept a list so that we can later pass in
lists of specific, user-requested search terms from the command-line.
But right now we still pass in the global search_terms list, so the
behavior of the program is the same.