Implement the --terms option
This command-line option allows the user to specific a list of terms
to search for in bookmarks.  This will override the global list of
search_terms and only use what is on the command-line.  Those terms do
not need to be part of the global search_terms list, meaning the user
can use this option to search for anything in their bookmarks.