[Feature] Implement --title-only
This patches adds a new `--title-only` argument which causes the
program to search only bookmark titles, not their URLs.  This is
useful when running something like:

    $ ./get-code-bookmarks.py $DB --terms git --title-only

Without `--title-only` this would show all bookmarks for pages on
sites like GitHub since the term appears in the `github.com` URL,
potentially creating a lot of false positives when trying to get
bookmarks that relate to Git itself.